Add api tests for targets and reuse datasets
